collapse trace and trace.intercept packages
