+++ spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/SpanHolder.java
@@ -1,0 +1,22 @@
+package org.springframework.cloud.sleuth;
+
+import lombok.extern.apachecommons.CommonsLog;
+
+/**
+ * @author Spencer Gibb
+ */
+@CommonsLog
+public class SpanHolder {
+	private static final ThreadLocal<Span> currentSpan = new ThreadLocal<>();
+
+	public static Span getCurrentSpan() {
+		return currentSpan.get();
+	}
+
+	public static void setCurrentSpan(Span span) {
+		if (log.isTraceEnabled()) {
+			log.trace("Setting current span " + span);
+		}
+		currentSpan.set(span);
+	}
+}
