+++ spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/SpanStartListener.java
@@ -1,0 +1,9 @@
+package org.springframework.cloud.sleuth;
+
+/**
+ * @author Spencer Gibb
+ */
+//TODO: rename?
+public interface SpanStartListener {
+	void startSpan(Span span);
+}
