+++ spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/TimelineAnnotation.java
@@ -1,0 +1,12 @@
+package org.springframework.cloud.sleuth;
+
+import lombok.Data;
+
+/**
+ * @author Spencer Gibb
+ */
+@Data
+public class TimelineAnnotation {
+	private final long time;
+	private final String msg;
+}
