+++ spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/TraceInfo.java
@@ -1,0 +1,12 @@
+package org.springframework.cloud.sleuth;
+
+import lombok.Data;
+
+/**
+ * @author Spencer Gibb
+ */
+@Data
+public class TraceInfo {
+	private final String traceId;
+	private final String spanId;
+}
