+++ spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/Utils.java
@@ -1,0 +1,14 @@
+package org.springframework.cloud.sleuth;
+
+import lombok.extern.apachecommons.CommonsLog;
+
+/**
+ * @author Spencer Gibb
+ */
+@CommonsLog
+public abstract class Utils {
+	public static void error(String msg) {
+		error(msg);
+		throw new RuntimeException(msg);
+	}
+}
