--- spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/trace/SpanStartListener.java
@@ -1,9 +1,0 @@
-package org.springframework.cloud.sleuth.trace;
-
-/**
- * @author Spencer Gibb
- */
-//TODO: rename?
-public interface SpanStartListener {
-	void startSpan(Span span);
-}
