--- spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/trace/TimelineAnnotation.java
@@ -1,12 +1,0 @@
-package org.springframework.cloud.sleuth.trace;
-
-import lombok.Data;
-
-/**
- * @author Spencer Gibb
- */
-@Data
-public class TimelineAnnotation {
-	private final long time;
-	private final String msg;
-}
