--- spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/trace/TraceAutoConfiguration.java
@@ -1,34 +1,0 @@
-package org.springframework.cloud.sleuth.trace;
-
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.cloud.sleuth.IdGenerator;
-import org.springframework.cloud.sleuth.RandomUuidGenerator;
-import org.springframework.cloud.sleuth.trace.sampler.IsTracingSampler;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import java.util.Collection;
-
-/**
- * @author Spencer Gibb
- */
-@Configuration
-public class TraceAutoConfiguration {
-
-	@Bean
-	@ConditionalOnMissingBean
-	public IdGenerator idGenerator() {
-		return new RandomUuidGenerator();
-	}
-
-	@Bean
-	public Sampler defaultSampler() {
-		return new IsTracingSampler();
-	}
-
-	@Bean
-	@ConditionalOnMissingBean
-	public Trace trace(Sampler sampler, IdGenerator idGenerator, Collection<SpanStartListener> listeners, Collection<SpanReceiver> receivers) {
-		return new DefaultTrace(sampler, idGenerator, listeners, receivers);
-	}
-}
