--- spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/trace/TraceInfo.java
@@ -1,12 +1,0 @@
-package org.springframework.cloud.sleuth.trace;
-
-import lombok.Data;
-
-/**
- * @author Spencer Gibb
- */
-@Data
-public class TraceInfo {
-	private final String traceId;
-	private final String spanId;
-}
