--- spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/trace/intercept/scheduling/TraceSchedulingAutoConfiguration.java
@@ -1,30 +1,0 @@
-package org.springframework.cloud.sleuth.trace.intercept.scheduling;
-
-/**
- * @author Spencer Gibb
- */
-
-import org.springframework.cloud.sleuth.trace.Trace;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-/**
- * Registers beans related to task scheduling.
- *
- * @see TraceSchedulingAspect
- *
- * @author Michal Chmielarz, 4financeIT
- * @author Spencer Gibb
- */
-@Configuration
-@EnableScheduling
-@EnableAspectJAutoProxy
-public class TraceSchedulingAutoConfiguration {
-
-	@Bean
-	public TraceSchedulingAspect traceSchedulingAspect(Trace trace) {
-		return new TraceSchedulingAspect(trace);
-	}
-}
