--- spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/trace/intercept/web/client/TraceWebClientAutoConfiguration.java
@@ -1,26 +1,0 @@
-package org.springframework.cloud.sleuth.trace.intercept.web.client;
-
-import org.springframework.boot.autoconfigure.AutoConfigureAfter;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.cloud.sleuth.resttemplate.SleuthRestTemplateAutoConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-/**
- * @author Spencer Gibb
- */
-@Configuration
-@ConditionalOnProperty(value = "spring.cloud.sleuth.trace.web.client.enabled", matchIfMissing = true)
-@ConditionalOnClass(RestTemplate.class)
-@AutoConfigureAfter(SleuthRestTemplateAutoConfiguration.class)
-public class TraceWebClientAutoConfiguration {
-
-	@Bean
-	@ConditionalOnMissingBean
-	public TraceRestTemplateInterceptor traceRestTemplateInterceptor() {
-		return new TraceRestTemplateInterceptor();
-	}
-}
