+++ spring-cloud-sleuth-core/src/main/resources/META-INF/spring.factories
@@ -1,7 +1,7 @@
 # Auto Configuration
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 org.springframework.cloud.sleuth.resttemplate.SleuthRestTemplateAutoConfiguration,\
-org.springframework.cloud.sleuth.trace.TraceAutoConfiguration,\
-org.springframework.cloud.sleuth.trace.intercept.web.TraceWebAutoConfiguration,\
-org.springframework.cloud.sleuth.trace.intercept.web.client.TraceWebClientAutoConfiguration,\
+org.springframework.cloud.sleuth.TraceAutoConfiguration,\
+org.springframework.cloud.sleuth.web.TraceWebAutoConfiguration,\
+org.springframework.cloud.sleuth.web.client.TraceWebClientAutoConfiguration,\
 org.springframework.cloud.sleuth.slf4j.SleuthSlf4jAutoConfiguration
